OcAudioLib: Fix compiler warning
closes acidanthera/bugtracker#738

+++ Library/OcAudioLib/OcAudioWave.c
@@ -164,7 +164,7 @@ InternalGetRawData (
       return EFI_UNSUPPORTED;
   }
 
-  *Channels = WaveFormat->Channels;
+  *Channels = (UINT8) WaveFormat->Channels;
 
   return EFI_SUCCESS;
 }
